lp-gateway: Add extrinsic for executing message recovery

+++ pallets/liquidity-pools-gateway/src/mock.rs
@@ -167,7 +167,7 @@ impl pallet_liquidity_pools_gateway::Config for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	type RuntimeOrigin = RuntimeOrigin;
 	type Sender = Sender;
-	type SessionId = u64;
+	type SessionId = u32;
 	type WeightInfo = ();
 }
